refactor [frontend] outsource Dialog to component
